sqrt_api/pyproject.toml
[tool.poetry]
name = "sqrt_api"
version = "0.0.1"
description = "Проект интеграции с купонатором Combo Mail.ru"
authors = ["Lensky Michael lenskymiwa@ya.ru"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = "^0.13.3"
uvloop = "^0.14.0"
PyYAML = "^5.3.1"
aiohttp = "^3.7.3"
httptools = "^0.1.1"
celery = "^5.0.5"
redis = "^3.5.3"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["setuptools", "poetry_core>=1.0"]
build-backend = "poetry.core.masonry.api"